feat: refactor PodcastsScreen to use LibraryScreen for improved structure and navigation
